Initial additional lyrics providers work
